Update: Changed Bar Graph Render
