+++ oil/Sources/Sections/Schemas/Graph.swift
@@ -8,6 +8,7 @@
 import Foundation
 import AppKit
 import Defaults
+import SwiftUI
 
 class Graph {
     internal var view: NSView = NSView(frame: .zero)
@@ -24,8 +25,19 @@ class Graph {
     deinit {
         view = NSView(frame: .zero)
     }
-    
-    func generateGraph(data: [Int]) {}
+
+    func generateGraph(data: [Int]) {
+        let shapeLayer: CAShapeLayer = CAShapeLayer()
+        shapeLayer.path = self.draw(data: data).cgPath
+        shapeLayer.fillColor = CGColor.white
+        //shapeLayer.strokeColor = CGColor.white
+        //shapeLayer.lineWidth = 0.5
+        if let oldLayer = view.layer?.sublayers?[0] {
+            view.layer?.replaceSublayer(oldLayer, with: shapeLayer)
+        } else {
+            view.layer?.addSublayer(shapeLayer)
+        }
+    }
     
     func resize() {
         self.stepSize = Defaults[.graphWidth]
@@ -35,5 +47,9 @@ class Graph {
                                          y: 0,
                                          width: self.stepSize * self.historyLength,
                                          height: 18))
+    }
+    
+    func draw(data: [Int]) -> Path {
+        return Path()
     }
 }
